+++ demos/rag/launch_query_test.py
@@ -1,6 +1,6 @@
 import json
 
-from demos.components.chroma_document_store import ChromaDocumentStore
+from demos.components.vectorstore.chroma_document_store import ChromaDocumentStore
 
 
 def pretty_print(json_obj):
